Remove SlashCSS framework
Nothing against it in particular (I didn't inspect it very thoroughly),
but...

- it's undocumented
- it's unmaintained
- it outputs far far more CSS than is actually used

This means we can remove a lot of dependencies (the whole Sass/CSS
toolchain), there's a whole lot less undocumented code for me to keep on
top of, and we save a huge amount of payload size. (*Minified* before:
211K; *unminified* now: 15K.)

I've kept the subset (thanks to uncss, plus a bit of manual processing)
which is actually used, unminified it, and put it in the static files.
There may be some unused or single-use custom properties still in there.
I kept some colours and sizes intentionally even though they're not
used.

Closes ML-95
Related to ML-128.
Part of ML-131.